moving test cases from proto_test.go to service_test.go
